Using pre-build instead of pre-push
making a pre-push hook to have the current commit hash added to the actual current commit hash will never work unless we have outside dependencies. Using pre-build instead, as it's included directly in STM32CubeIIDE. And when we build, we are grabbing the current commit. So, the current commit just won't be reflected upon pushing.